Prevent missing directory from failing test
Summary: If the directory doesn't exist then don't try to remove it.

Test Plan: all tests pass, deleted runfiles dir to ensure.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh

Maniphest Tasks: T17458

Differential Revision: https://phabricator.sourcevertex.net/D24015

+++ tensorflow/python/ipu/tests/keras/keras_pipeline_save_restore_test.py
@@ -133,7 +133,8 @@ class IPUPipelineTest(test.TestCase):
     checkpoint_dir = os.path.dirname(checkpoint_path)
 
     # Remove directory from any previous run
-    shutil.rmtree(checkpoint_dir)
+    if os.path.isdir(checkpoint_dir):
+      shutil.rmtree(checkpoint_dir)
 
     strategy = ipu.ipu_strategy.IPUStrategy()
     with strategy.scope():
